separated simple pyFields and pyFields with mappings
