+++ src/pyHiChi/CMakeLists.txt
@@ -21,6 +21,7 @@ endif()
 
 set(pyHiChi_source
     include/pyField.h
+    include/pyFieldInterface.h
     src/pyHiChi.cpp)
 
 pybind11_add_module(pyHiChi ${pyHiChi_source})
